splitting tests for future maintainibility
