+++ Assets/Scripts/Tests/EditMode/QueenTests.cs
@@ -1,17 +1,13 @@
 using System.Collections;
+using Assets.Scripts.Classes.Pieces;
 using NUnit.Framework;
 using UnityEngine.TestTools;
 
 namespace Assets.Scripts.Tests.EditMode
 {
-    public class PenetrationTests
+    public class QueenTests
     {
-        // A Test behaves as an ordinary method
-        [Test]
-        public void PenetrationTestsSimplePasses()
-        {
-            // Use the Assert class to test conditions
-        }
+        
 
         // A UnityTest behaves like a coroutine in Play Mode. In Edit Mode you can use
         // `yield return null;` to skip a frame.
