+++ Assets/Scripts/Tests/EditMode/Tests.asmdef
@@ -5,5 +5,6 @@
     ],
     "includePlatforms": [
         "Editor"
-    ]
+    ],
+    "references":[ "GUID:74413e586c74caa409add786cbb74a70" ]
 }
