+++ Assets/Scripts/Tests/PlayMode/PieceValidationTests.cs
@@ -1,0 +1,37 @@
+﻿using System;
+using System.Collections;
+using System.Collections.Generic;
+using Assets.Scripts.Classes.GameClasses;
+using Assets.Scripts.Classes.PieceComponent;
+using NUnit.Framework;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+using UnityEngine.TestTools;
+using static Assets.Scripts.Utility;
+
+namespace Tests.PlayMode
+{
+    public class PieceValidationTests
+    {
+        [UnityTest]
+        public IEnumerator TestPawnCapturing()
+        {
+            SceneManager.LoadScene("GameScene");
+            yield return new WaitUntil(() => SceneManager.GetActiveScene().name == "GameScene");
+            yield return new WaitForSeconds(1.1f);
+            Dictionary<Vector2Int, PieceMovementComponent> pieces = GameManager.Instance.Pieces;
+            var e4Pawn = pieces[new Vector2Int(5, 2)] ?? throw new ArgumentNullException("pieces[new Vector2Int(5, 2)]"); // Pawn At (5.2)
+            Assert.IsNotNull(e4Pawn);
+            var move1 = PawnValidator(pieces, (Vector2Int)e4Pawn.CurrPos, new Vector2Int(6, 3));
+            Assert.AreEqual(move1, false);
+            move1=PawnValidator(pieces, (Vector2Int)e4Pawn.CurrPos, new Vector2Int(4, 3));
+            Assert.AreEqual(move1, false);
+            var e7Pawn = pieces[new Vector2Int(5, 7)];
+            Assert.IsNotNull(e7Pawn);
+            move1 = PawnValidator(pieces, (Vector2Int)e7Pawn.CurrPos, new Vector2Int(6, 6));
+            Assert.AreEqual(move1, false);
+            move1 = PawnValidator(pieces, (Vector2Int)e7Pawn.CurrPos, new Vector2Int(4, 6));
+            Assert.AreEqual(move1, false);
+        }
+    }
+}
